fix warning: cast -1 to uint
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@1353 3db68df2-c116-0410-a063-a993310a9797

+++ source/scripting/DOMEvent.h
@@ -52,7 +52,7 @@ public:
 	jsval PreventDefault( JSContext* cx, uintN argc, jsval* argv );
 
 public:
-	CScriptEvent( const CStrW& Type, bool Cancelable, unsigned int TypeCode = -1 );
+	CScriptEvent( const CStrW& Type, bool Cancelable, unsigned int TypeCode = (unsigned int)-1 );
 	static void ScriptingInit();
 };
